Updates to propeller locations and updates to the lee_controller_f450.yaml as well some other minor updates
Unfortunately Brennan and I are having trouble with making the transition from rotors_... to f450_... . We know this is something we can figure out, but it just is taking more time than we were hoping for. Also, for the lee_controller_f450.yaml, I am seeing more results with how I have tried this new approach that Dr. Vaughan sent over to me today. It appears that I am seeing more of the trend that is letting me control the drone much better